+++ src/FlowExplainer.Desktop/Fields/AnalyticalEvolvingVelocityField.cs
@@ -1,16 +1,16 @@
 using ImGuiNET;
-using static System.Single;
+using static System.Double;
 
 namespace FlowExplainer;
 
 //https://shaddenlab.berkeley.edu/uploads/LCS-tutorial/examples.html#x1-1200812
 public class AnalyticalEvolvingVelocityField : IVectorField<Vec3, Vec2>
 {
-    public float epsilon = .1f;
-    public float A = 1f;
-    public float w = 0.002f;
+    public double epsilon = .1f;
+    public double A = 1f;
+    public double w = 0.002f;
 
-    public float Period => (2f * Pi) / w;
+    public double Period => (2f * Pi) / w;
     public IDomain<Vec3> Domain => new RectDomain<Vec3>(new Vec3(0, 0, 0), new Vec3(2, 1f, Period), BoundingFunctions.None<Vec3>());
 
     public Vec3 Wrap(Vec3 x)
@@ -34,25 +34,25 @@ public class AnalyticalEvolvingVelocityField : IVectorField<Vec3, Vec2>
         ImGuiHelpers.SliderFloat("w", ref w, 0, 2);
     }
 
-    float streamFunction(Vec3 phase)
+    double streamFunction(Vec3 phase)
     {
-        float x = phase.X;
-        float y = phase.Y;
-        float t = phase.Z;
+        double x = phase.X;
+        double y = phase.Y;
+        double t = phase.Z;
         return A * Sin(Pi * f(x, t)) * Sin(Pi * y);
     }
 
-    float a(float t)
+    double a(double t)
     {
         return epsilon * Sin(w * t);
     }
 
-    float b(float t)
+    double b(double t)
     {
         return 1f - 2 * epsilon * Sin(w * t);
     }
 
-    public float f(float x, float t)
+    public double f(double x, double t)
     {
         return a(t) * x * x + b(t) * x;
     }
@@ -70,7 +70,7 @@ public class AnalyticalEvolvingVelocityField : IVectorField<Vec3, Vec2>
         return Velocity(x.X, x.Y, x.Z);
     }
 
-    public Vec2 Velocity(float x, float y, float t)
+    public Vec2 Velocity(double x, double y, double t)
     {
         var u = -Pi * A * Sin(Pi * f(x, t)) * Cos(Pi * y);
         var dfdx = 2 * a(t) * x + b(t);
